extract resused code into functions
