+++ typescript/infra/scripts/xerc20/xerc20vs-set-limits.ts
@@ -8,12 +8,8 @@ import {
 } from '@hyperlane-xyz/utils';
 
 import {
-  getRegistry,
-  getWarpAddresses,
-  getWarpCoreConfig,
-} from '../../config/registry.js';
-import {
   deriveBridgesConfig,
+  getWarpConfigsAndArtifacts,
   updateChainLimits,
 } from '../../src/xerc20/utils.js';
 import { getArgs, withWarpRouteIdRequired } from '../agent-utils.js';
@@ -24,17 +20,8 @@ async function main() {
   const { environment, warpRouteId } = await withWarpRouteIdRequired(getArgs())
     .argv;
 
-  const registry = getRegistry();
-  const warpDeployConfig = registry.getWarpDeployConfig(warpRouteId);
-  const warpCoreConfig = registry.getWarpRoute(warpRouteId);
-  if (!warpDeployConfig) {
-    throw new Error(`Warp deploy config for route ID ${warpRouteId} not found`);
-  }
-  if (!warpCoreConfig) {
-    throw new Error(`Warp core config for route ID ${warpRouteId} not found`);
-  }
-
-  const warpAddresses = getWarpAddresses(warpRouteId);
+  const { warpDeployConfig, warpCoreConfig, warpAddresses } =
+    getWarpConfigsAndArtifacts(warpRouteId);
 
   const envConfig = getEnvironmentConfig(environment);
   const multiProtocolProvider = await envConfig.getMultiProtocolProvider();
